+++ arrakis/rmsynth_oncuts.py
@@ -9,8 +9,9 @@ import warnings
 from pathlib import Path
 from pprint import pformat
 from shutil import copyfile
-from typing import List, Optional, Tuple, Union
+from typing import List
 from typing import NamedTuple as Struct
+from typing import Optional, Tuple, Union
 
 import astropy.units as u
 import matplotlib
